+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.9] - 2021-04-13
+## Added
+- Option enabling for the user to retry when the access to the camera failed
+
+## Fixed
+- Camera error notification also displayed in case of 1+ access to the camera
+
 ## [1.1.8] - 2021-04-09
 ### Added
 - Documentation for contributions and deployment process